Fixing styles inside h4 used in content body
Currently, this h4~* declaration affects everything next to it.

See the note div here:
https://success.outsystems.com/Documentation/11/Developing_an_Application/Design_UI/Accessibility#Change_the_role_of_the_Alert_pattern

I believe this declaration is used to address the mobile version of the documentation/site menu, specifically the "Go back" bullet/icon. I tried to restrict that to the header, which I believe should suffice, but would appreciate the code review @diegoalmeida as you know this better than we do.

Thanks.